# back/docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /adventurers:
    get:
      operationId: AdventurersController_findAll
      parameters:
        - name: name
          required: false
          in: query
          description: Filtrer par nom (contains, insensible à la casse)
          schema:
            example: aria
            type: string
        - name: specialityId
          required: false
          in: query
          description: Filtrer par identifiant de spécialité
          schema:
            example: 3
            type: number
        - name: experienceMin
          required: false
          in: query
          description: Expérience minimale (incluse)
          schema:
            example: 10
            type: number
        - name: experienceMax
          required: false
          in: query
          description: Expérience maximale (incluse)
          schema:
            example: 50
            type: number
        - name: dailyRateOrder
          required: false
          in: query
          description: Tri par taux journalier
          schema:
            example: asc
            type: string
            enum:
              - asc
              - desc
      responses:
        '200':
          description: List of adventurers (avec filtres et tri)
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: number
                      example: 1
                    name:
                      type: string
                      example: Aria Stormblade
                    specialityId:
                      type: number
                      example: 3
                    dailyRate:
                      type: number
                      example: 150
                    experience:
                      type: number
                      example: 120
                    availableUntil:
                      type: string
                      format: date-time
                      nullable: true
                      example: '2025-12-31T23:59:59.000Z'
      security:
        - bearer: []
      tags:
        - Adventurers
    post:
      operationId: AdventurersController_create
      parameters: []
      requestBody:
        required: true
        description: New adventurer payload
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Aria Stormblade
                specialityId:
                  type: number
                  example: 3
                dailyRate:
                  type: number
                  example: 100
                equipmentTypeIds:
                  type: array
                  items:
                    type: number
                  example:
                    - 1
                    - 2
                consumableTypeIds:
                  type: array
                  items:
                    type: number
                  example:
                    - 1
              required:
                - name
                - specialityId
                - dailyRate
      responses:
        '201':
          description: Adventurer created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 42
                  name:
                    type: string
                    example: Aria Stormblade
                  specialityId:
                    type: number
                    example: 3
                  dailyRate:
                    type: number
                    example: 100
                  experience:
                    type: number
                    example: 0
                  availableUntil:
                    type: string
                    format: date-time
                    nullable: true
                    example: null
      security:
        - bearer: []
      tags:
        - Adventurers
  /adventurers/{id}:
    get:
      operationId: AdventurersController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: Adventurer ID
          schema:
            example: 1
            type: number
      responses:
        '200':
          description: Adventurer by id
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 1
                  name:
                    type: string
                    example: Aria Stormblade
                  specialityId:
                    type: number
                    example: 3
                  dailyRate:
                    type: number
                    example: 150
                  experience:
                    type: number
                    example: 120
                  availableUntil:
                    type: string
                    format: date-time
                    nullable: true
                    example: '2025-12-31T23:59:59.000Z'
      security:
        - bearer: []
      tags:
        - Adventurers
    patch:
      operationId: AdventurersController_update
      parameters:
        - name: id
          required: true
          in: path
          description: Adventurer ID
          schema:
            example: 42
            type: number
      requestBody:
        required: true
        description: Fields to update (partial)
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Aria Nightwind
                dailyRate:
                  type: number
                  example: 180
                specialityId:
                  type: number
                  example: 2
                equipmentTypeIds:
                  type: array
                  items:
                    type: number
                  example:
                    - 1
                    - 3
                consumableTypeIds:
                  type: array
                  items:
                    type: number
                  example:
                    - 2
              additionalProperties: false
      responses:
        '200':
          description: Updated adventurer
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 42
                  name:
                    type: string
                    example: Aria Nightwind
                  specialityId:
                    type: number
                    example: 2
                  dailyRate:
                    type: number
                    example: 180
                  experience:
                    type: number
                    example: 120
                  availableUntil:
                    type: string
                    format: date-time
                    nullable: true
                    example: '2025-12-31T23:59:59.000Z'
      security:
        - bearer: []
      tags:
        - Adventurers
...
info:
  title: API "Les kips AH"
  description: |-
    This is the API information regarding the project "Les capuches d'Opale".

    C'est un outil de gestion d'aventurier pour une guilde.
  version: 1.0.0
  contact: {}
  termsOfService: https://swagger.io/terms/
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http